+++ src/QNEthernet.cpp
@@ -17,7 +17,6 @@
 #include "lwip/dns.h"
 #include "lwip/igmp.h"
 #include "lwip/ip_addr.h"
-#include "lwip/opt.h"
 #include "lwip/udp.h"
 #include "lwip_t41.h"